refactor: Move commands module into main.rs
